FileAutomationGUI/requirements.txt
altgraph==0.17.2
future==0.18.2
lxml==4.7.1
pefile==2021.9.3
Pillow==9.0.0
pydub==0.25.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.5
PyPDF2==1.26.0
python-docx==0.8.11
pywin32==303
pywin32-ctypes==0.2.0
